+++ client-core/src/main/java/com/influxdb/query/internal/FluxResultMapper.java
@@ -25,6 +25,7 @@ import java.lang.reflect.Field;
 import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Objects;
+import java.util.regex.Pattern;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
@@ -33,6 +34,9 @@ import com.influxdb.exceptions.InfluxException;
 import com.influxdb.query.FluxRecord;
 
 public class FluxResultMapper {
+
+    // we don't cover some corner cases(e.g. someURI someHRName) under which explicit mapping is required
+    private static final Pattern CAMEL_CASE_TO_SNAKE_CASE_REPLACE_PATTERN = Pattern.compile("[A-Z]");
 
     /**
      * Maps FluxRecord into custom POJO class.
@@ -71,6 +75,11 @@ public class FluxResultMapper {
                         col = columnName;
                     } else if (recordValues.containsKey("_" + columnName)) {
                         col = "_" + columnName;
+                    } else {
+                        String columnNameInSnakeCase = camelCaseToSnakeCase(columnName);
+                        if (recordValues.containsKey(columnNameInSnakeCase)) {
+                            col = columnNameInSnakeCase;
+                        }
                     }
 
                     if (col != null) {
@@ -86,6 +95,12 @@ public class FluxResultMapper {
         } catch (Exception e) {
             throw new InfluxException(e);
         }
+    }
+
+    private String camelCaseToSnakeCase(final String str) {
+        return CAMEL_CASE_TO_SNAKE_CASE_REPLACE_PATTERN.matcher(str)
+            .replaceAll("_$0")
+            .toLowerCase();
     }
 
     private void setFieldValue(@Nonnull final Object object,